feat(GOG-44): :recycle: refactor utils file into smaller chunks
Also better document some util functions
